models path refactored
